BIG UGLY COMMIT; restructuring and -naming
All the rest should be retrievable from the AoC website.
It's possible that it'll go offline or disappear, but I'll most likely
be dead by then, or not interested anymore due to senility or some other
old man's condition :-)

And it will of course also be handier when using the simplified `aoc`
command in my dotfiles.